+++ tests/test_loaders.py
@@ -3,11 +3,14 @@ from unittest import TestCase
 
 from liquid import BoundTemplate, Context, Environment, FileExtensionLoader
 from liquid.exceptions import TemplateNotFound
-from liquid.loaders import BaseLoader
+from liquid.loaders import BaseLoader, PackageLoader
 from pytest import raises
 
-from fhir_converter.loaders import ResourceLoader, TemplateSystemLoader, read_text
+from fhir_converter.loaders import make_template_system_loader, read_text
 from fhir_converter.tags import all_tags, register_tags
+
+user_defined_ccda_loader = FileExtensionLoader(search_path="data/templates/ccda")
+ccda_loader = PackageLoader(package="fhir_converter.templates", package_path="ccda")
 
 
 def build_env(loader: BaseLoader, **kwargs) -> Environment:
@@ -16,76 +19,16 @@ def build_env(loader: BaseLoader, **kwargs) -> Environment:
     return env
 
 
-class ResourceLoaderTest(TestCase):
-    @staticmethod
-    def get_rendering_env(
-        search_package: str = "fhir_converter.templates.ccda", **kwargs
-    ) -> Environment:
-        return build_env(ResourceLoader(search_package), **kwargs)
-
-    def test_load_template(self) -> None:
-        template = self.get_rendering_env().get_template(name="CCD.liquid")
-        self.assertIsInstance(template, BoundTemplate)
-
-    def test_load_template_relative(self) -> None:
-        template = self.get_rendering_env().get_template(name="./CCD.liquid")
-        self.assertIsInstance(template, BoundTemplate)
-
-    def test_load_template_without_suffix(self) -> None:
-        template = self.get_rendering_env().get_template(name="CCD")
-        self.assertIsInstance(template, BoundTemplate)
-
-    def test_cached_template(self) -> None:
-        env = self.get_rendering_env()
-        template = env.get_template(name="CCD.liquid")
-        self.assertIsInstance(template, BoundTemplate)
-
-        another = env.get_template(name="CCD.liquid")
-        self.assertIsInstance(template, BoundTemplate)
-        self.assertEqual(template.tree, another.tree)
-
-    def test_disable_template_cache(self) -> None:
-        env = self.get_rendering_env(cache_size=0)
-        template = env.get_template(name="CCD.liquid")
-        self.assertTrue(template.is_up_to_date)
-
-        another = env.get_template(name="CCD.liquid")
-        self.assertTrue(another.is_up_to_date)
-        self.assertIsNot(template.tree, another.tree)
-
-    def test_template_not_found(self) -> None:
-        with raises(TemplateNotFound):
-            self.get_rendering_env().get_template(name="nosuchthing.liquid")
-
-    def test_no_such_search_path(self) -> None:
-        with raises(TemplateNotFound):
-            self.get_rendering_env(search_package="nosuchthing").get_template(
-                name="nosuchthing.liquid"
-            )
-
-    def test_escape_search_path(self) -> None:
-        with raises(TemplateNotFound):
-            self.get_rendering_env().get_template(name="../CCD.liquid")
-
-    def test_escape_search_path_nested(self) -> None:
-        with raises(TemplateNotFound):
-            self.get_rendering_env().get_template(name="./../CCD.liquid")
-
-
 class TemplateSystemLoaderTest(TestCase):
-    loaders = [
-        FileExtensionLoader(search_path="data/templates/ccda"),
-        ResourceLoader(search_package="fhir_converter.templates.ccda"),
-    ]
-
     def get_rendering_env(
         self,
         loader_cache_size: int = 1,
         **kwargs,
     ) -> Environment:
         return build_env(
-            TemplateSystemLoader(
-                loaders=self.loaders,
+            make_template_system_loader(
+                loader=user_defined_ccda_loader,
+                additional_loaders=[ccda_loader],
                 cache_size=loader_cache_size,
             ),
             **kwargs,
@@ -122,7 +65,7 @@ class TemplateSystemLoaderTest(TestCase):
         template = self.get_rendering_env().get_template(name="Utils/GenerateId")
         self.assertIsInstance(template, BoundTemplate)
 
-    def test_cached_template(self) -> None:
+    def test_cached_user_defined_template(self) -> None:
         env = self.get_rendering_env()
         template = env.get_template(name="Pampi.liquid")
         self.assertIsInstance(template, BoundTemplate)
@@ -140,7 +83,7 @@ class TemplateSystemLoaderTest(TestCase):
         self.assertEqual({"test": "two"}, template.globals)
         self.assertEqual(template.tree, another.tree)
 
-    def test_cached_resource_template(self) -> None:
+    def test_cached_package_template(self) -> None:
         env = self.get_rendering_env()
         template = env.get_template(name="CCD.liquid")
         self.assertIsInstance(template, BoundTemplate)
@@ -160,18 +103,16 @@ class TemplateSystemLoaderTest(TestCase):
         self.assertTrue(another.is_up_to_date)
         self.assertEqual(template.tree, another.tree)
 
-    def test_disable_loader_cache_template_not_cached(self) -> None:
+    def test_disable_loader_cache_template_cached(self) -> None:
         env = self.get_rendering_env(loader_cache_size=0)
         template = env.get_template(name="Pampi.liquid")
         self.assertTrue(template.is_up_to_date)
 
-        # The environment only caches when the loader is not a caching loader,
-        # if this changes, this will not be None
-        self.assertIsNone(env.cache)
+        self.assertIsNotNone(env.cache)
 
         another = env.get_template(name="Pampi.liquid")
         self.assertTrue(another.is_up_to_date)
-        self.assertIsNot(template.tree, another.tree)
+        self.assertEqual(template.tree, another.tree)
 
     def test_disable_caches_template_not_cached(self) -> None:
         env = self.get_rendering_env(loader_cache_size=0, cache_size=0)
@@ -188,7 +129,7 @@ class TemplateSystemLoaderTest(TestCase):
 
 
 class ReadTextTest(TestCase):
-    env = build_env(ResourceLoader(search_package="fhir_converter.templates.ccda"))
+    env = build_env(ccda_loader)
 
     def test_file_not_found(self) -> None:
         with raises(FileNotFoundError, match="File not found nosuchthing.json"):