+++ tests/test_renderers.py
@@ -2,6 +2,7 @@ from collections.abc import Mapping
 from io import StringIO
 from json import loads
 from pathlib import Path
+from typing import cast
 from unittest import TestCase
 
 from liquid import FileExtensionLoader
@@ -10,57 +11,56 @@ from pyexpat import ExpatError
 from pyjson5 import Json5EOF
 from pytest import raises
 
-from fhir_converter.loaders import TemplateSystemLoader
+from fhir_converter.loaders import CachingTemplateSystemLoader, TemplateSystemLoader
 from fhir_converter.renderers import (
     CcdaRenderer,
     RenderingError,
     Stu3FhirRenderer,
     ccda_default_loader,
-    get_environment,
+    make_environment,
     stu3_default_loader,
 )
 
 
-class GetEnvironmentTest(TestCase):
+class MakeEnvironmentTest(TestCase):
     def test_defaults(self) -> None:
-        env = get_environment(loader=ccda_default_loader)
+        env = make_environment(loader=ccda_default_loader)
         self.assertFalse(env.auto_reload)
         self.assertIsNone(env.cache)
 
-        self.assertIsInstance(env.loader, TemplateSystemLoader)
-        loader: TemplateSystemLoader = env.loader  # type: ignore
+        self.assertIsInstance(env.loader, CachingTemplateSystemLoader)
+        loader = cast(CachingTemplateSystemLoader, env.loader)
         self.assertFalse(loader.auto_reload)
-        self.assertTrue(loader.is_caching)
         self.assertEqual([ccda_default_loader], loader.loaders)
 
         self.assertIsInstance(loader.cache, LRUCache)
         self.assertEqual(loader.cache.capacity, 300)  # type: ignore
 
     def test_auto_reload(self) -> None:
-        env = get_environment(loader=ccda_default_loader, auto_reload=True)
+        env = make_environment(loader=ccda_default_loader, auto_reload=True)
         self.assertFalse(env.auto_reload)
         self.assertIsNone(env.cache)
 
-        self.assertTrue(env.loader.auto_reload)  # type: ignore
+        loader = cast(CachingTemplateSystemLoader, env.loader)
+        self.assertTrue(loader.auto_reload)
 
     def test_cache_size(self) -> None:
-        env = get_environment(loader=ccda_default_loader, cache_size=1)
+        env = make_environment(loader=ccda_default_loader, cache_size=1)
         self.assertFalse(env.auto_reload)
         self.assertIsNone(env.cache)
 
-        self.assertTrue(env.loader.is_caching)  # type: ignore
-        self.assertEqual(env.loader.cache.capacity, 1)  # type: ignore
+        loader = cast(CachingTemplateSystemLoader, env.loader)
+        self.assertEqual(loader.cache.capacity, 1)  # type: ignore
 
     def test_cache_disabled(self) -> None:
-        env = get_environment(loader=ccda_default_loader, cache_size=0)
+        env = make_environment(loader=ccda_default_loader, cache_size=0)
         self.assertFalse(env.auto_reload)
         self.assertIsNone(env.cache)
-        self.assertFalse(env.loader.is_caching)  # type: ignore
-        self.assertEqual(env.loader.cache.capacity, 0)  # type: ignore
+        self.assertIsInstance(env.loader, TemplateSystemLoader)
 
     def test_additional_loaders(self) -> None:
         loader = FileExtensionLoader(search_path="data/templates/ccda")
-        env = get_environment(
+        env = make_environment(
             loader,
             additional_loaders=[ccda_default_loader],
         )
@@ -160,7 +160,7 @@ class Stu3FhirRendererTest(TestCase):
         self.validate_str(buffer.getvalue())
 
     def test_render_env_provided(self) -> None:
-        renderer = Stu3FhirRenderer(get_environment(loader=stu3_default_loader))
+        renderer = Stu3FhirRenderer(make_environment(loader=stu3_default_loader))
         self.validate(
             renderer.render_to_fhir(
                 "Immunization", self.stu3_file.read_text(encoding="utf-8")
@@ -254,7 +254,7 @@ class CcdaRendererTest(TestCase):
         self.validate_str(buffer.getvalue())
 
     def test_render_env_provided(self) -> None:
-        renderer = CcdaRenderer(get_environment(loader=ccda_default_loader))
+        renderer = CcdaRenderer(make_environment(loader=ccda_default_loader))
         self.validate(
             renderer.render_to_fhir("CCD", self.ccda_file.read_text(encoding="utf-8"))
         )